fz.almond_click
